+++ @planetarium/tx/src/tx/metadata.ts
@@ -1,4 +1,9 @@
-import { Encodable } from "bencodex";
+import {
+  BencodexDictionary,
+  Dictionary,
+  Key,
+  Value,
+} from "@planetarium/bencodex";
 import { Address, encodeAddress, encodeAddressSet } from "../address.js";
 import { BlockHash, encodeBlockHash } from "../blockhash.js";
 import { encodePublicKey, PublicKey } from "../key.js";
@@ -10,34 +15,31 @@ import { encodePublicKey, PublicKey } from "../key.js";
 export interface TxMetadata {
   nonce: bigint;
   publicKey: PublicKey;
-  signer: Address;  // TODO: This field can be derived from publicKey.
+  signer: Address; // TODO: This field can be derived from publicKey.
   timestamp: Date;
   updatedAddresses: Set<Address>;
   genesisHash: BlockHash | null;
 }
 
-const NONCE_KEY = Buffer.from([0x6e]); // 'n'
-const SIGNER_KEY = Buffer.from([0x73]); // 's'
-const GENESIS_HASH_KEY = Buffer.from([0x67]); // 'g'
-const UPDATED_ADDRESSES_KEY = Buffer.from([0x75]); // 'u'
-const PUBLIC_KEY_KEY = Buffer.from([0x70]); // 'p'
-const TIMESTAMP_KEY = Buffer.from([0x74]); // 't'
+const NONCE_KEY = new Uint8Array([0x6e]); // 'n'
+const SIGNER_KEY = new Uint8Array([0x73]); // 's'
+const GENESIS_HASH_KEY = new Uint8Array([0x67]); // 'g'
+const UPDATED_ADDRESSES_KEY = new Uint8Array([0x75]); // 'u'
+const PUBLIC_KEY_KEY = new Uint8Array([0x70]); // 'p'
+const TIMESTAMP_KEY = new Uint8Array([0x74]); // 't'
 
-export function encodeTxMetadata(
-  metadata: TxMetadata
-): Map<string | Buffer,Encodable> {
+export function encodeTxMetadata(metadata: TxMetadata): Dictionary {
   const updatedAddresses = encodeAddressSet(metadata.updatedAddresses);
-  const timestamp = metadata.timestamp.toISOString()
-    .replace(/Z$/, "000Z");
-  const dict = new Map<string | Buffer, Encodable>([
+  const timestamp = metadata.timestamp.toISOString().replace(/Z$/, "000Z");
+  const pairs: [Key, Value][] = [
     [NONCE_KEY, metadata.nonce],
     [SIGNER_KEY, encodeAddress(metadata.signer)],
     [UPDATED_ADDRESSES_KEY, updatedAddresses],
     [PUBLIC_KEY_KEY, encodePublicKey(metadata.publicKey)],
     [TIMESTAMP_KEY, timestamp],
-  ]);
+  ];
   if (metadata.genesisHash !== null) {
-    dict.set(GENESIS_HASH_KEY, encodeBlockHash(metadata.genesisHash));
+    pairs.push([GENESIS_HASH_KEY, encodeBlockHash(metadata.genesisHash)]);
   }
-  return dict;
+  return new BencodexDictionary(pairs);
 }