+++ @planetarium/tx/test/key.test.ts
@@ -5,13 +5,10 @@ import { bytesEqual } from "../src/bytes";
 
 test("encodePublicKey", () => {
   fc.assert(
-    fc.property(
-      fc.uint8Array({ minLength: 1 }),
-      (bytes: Uint8Array) => {
-        const key = encodePublicKey(bytes);
-        return key instanceof ArrayBuffer && bytesEqual(key, bytes);
-      }
-    )
+    fc.property(fc.uint8Array({ minLength: 1 }), (bytes: Uint8Array) => {
+      const key = encodePublicKey(bytes);
+      return key instanceof Uint8Array && bytesEqual(key, bytes);
+    }),
   );
   expect(() => encodePublicKey(new Uint8Array(0))).throws(TypeError, /empty/i);
 });